refactor(tests): move public-API put tests from inline mod to tests/cli_put_*.rs
Migrates 20 tests out of `src/commands/put.rs::mod tests` into 5
per-feature integration test files under `tests/`, leaving 23
white-box tests inline with a module-level annotation listing the
private items they exercise. Pure mechanical refactor: no production
code, visibility, or test logic changes — only test relocation with
adjusted `use` paths.

New files (each ≤ 1500-LOC budget):
- tests/cli_put_create.rs (3 tests) — first-write semantics: version 1, wing derivation, default page type
- tests/cli_put_update_occ.rs (5 tests + 1 cfg(not(unix))) — optimistic-concurrency update semantics + raw_imports rotation invariant
- tests/cli_put_supersede.rs (3 tests) — supersede chain linkage and rejection of non-head supersede
- tests/cli_put_source_invariants.rs (4 tests) — source-text invariants on writer ordering, IPC peer-auth seam, dedup fail-closed seam
- tests/cli_put_render.rs (4 tests) — round-trip render, quaid_id retention, JSON frontmatter, FTS5 indexing trigger
- tests/common/put_fixtures.rs — `open_test_db`, `read_page`, `superseded_by_for_slug`, `page_id_for_slug`, and `active_raw_import_*` helpers (each used by ≥ 2 of the new files)

Inline residue (23 tests, 1325 LOC):
private items exercised include `PutTestHooks`, `WriteLockBlockState`,
`SupersedeClaimBlockState`, `put_test_hooks`,
`write_lock_blocker_for_path`, `supersede_claim_blocker_for_path`,
`put_from_cli_string`, `write_dedup_key`, `sha256_hex`, and
`vault_sync::has_write_dedup` (cfg(test)-gated).

Verification:
  baseline (parent f0e623e):  passed=1357 failed=0
  this commit:                passed=1357 failed=0

Both clippy channels green:
  cargo clippy --all-targets --locked -- -D warnings
  cargo clippy --all-targets --no-default-features \
    --features bundled,online-model --locked -- -D warnings

src/commands/put.rs:  3268 LOC -> 2718 LOC (-550 LOC, -17%)

Part of OpenSpec change `extract-inline-tests-to-integration`.
